DXBE-20: Add source:config:pull, extract shared Source command base
Adds source:config:pull, a mirror of push that triggers a SAS config
export (CMS to repo). Extracts the shared trigger/poll flow into an
abstract ConfigCommandBase so both directions reuse the same SAS client
wiring; each subclass supplies only its endpoint call and messaging.
Generalizes SourceConfig::getStatus() to serve both operations.
